refactor[#345] : 마킹 좋아요 / 좋아요 취소 api 요청 함수에 apiClient 적용
