+++ src/features/marking/api/index.ts
@@ -2,5 +2,6 @@ export * from "./getMarkingList";
 export * from "./postMarkingForm";
 export * from "./postTempMarkingForm";
 export * from "./deleteMarking";
-export * from "./likeMarking";
+export * from "./postLikeMarking";
+export * from "./deleteLikeMarking";
 export * from "./saveMarking";
